+++ EmergencyCallServer/utils/GeoPoint.cs
@@ -2,10 +2,10 @@
 {
     public class GeoPoint
     {
-        public double Latitude { get; }
-        public double Longitude { get; }
+        public decimal Latitude { get; }
+        public decimal Longitude { get; }
 
-        public GeoPoint(double latitude, double longitude)
+        public GeoPoint(decimal latitude, decimal longitude)
         {
             Latitude = latitude;
             Longitude = longitude;
